Bug Fix: smsdebug.c --> local buffer used to store msg is now made global from local; this was causing stack overflow
Modification:

main.c:
--. Few minor addition 

main.h:
--> http states are added 

http_file.c:
--> upload file FSM is made
--> states are defined now for uploading file

modem.h:
--> pin port as RST is configured but not yet used
--> sw_rts is added in modem structure

Modem/ModemInterface.c:
--> Few changes in TCP and UDP connect

+++ CooCox/STM32GatwayRev0.1/Modem/http.h
@@ -2,8 +2,8 @@
  * Decleration of the Public functions
  */
 uint8_t uploadFile(mdmIface *mdm, const char *file, server *tcp);
-httpStatus mdmHttpRes(mdmIface *mdm, uint16_t* ContLen, uint8_t cond);
-httpStatus mdmHttpBody(mdmIface *mdm, const char* resp, uint16_t *len, uint16_t timeout);
+httpStatus mdmHttpRes(mdmIface *mdm, uint32_t* ContLen, uint8_t cond);
+httpStatus mdmHttpBody(mdmIface *mdm, const char* resp, uint32_t *len, uint16_t timeout);
 /*
  * Decleration of the Private functions
  */
